requirements.txt
jupyter
matplotlib
numpy
pandas
mlflow
boto3
minio
python-dotenv